refactored ObjectProposal layer
