+++ keras_rcnn/backend/tensorflow_backend.py
@@ -1,8 +1,6 @@
 import keras.backend
 import tensorflow
 
-import keras_rcnn.backend
-
 RPN_NEGATIVE_OVERLAP = 0.3
 RPN_POSITIVE_OVERLAP = 0.7
 RPN_FG_FRACTION = 0.5
@@ -96,36 +94,6 @@ def non_maximum_suppression(boxes, scores, maximum, threshold=0.5):
         max_output_size=maximum,
         scores=scores
     )
-
-
-def propose(boxes, scores, maximum):
-    shape = keras.backend.shape(boxes)[1:3]
-
-    shifted = keras_rcnn.backend.shift(shape, 16)
-
-    proposals = keras.backend.reshape(boxes, (-1, 4))
-
-    proposals = keras_rcnn.backend.bbox_transform_inv(shifted, proposals)
-
-    proposals = keras_rcnn.backend.clip(proposals, shape)
-
-    indicies = keras_rcnn.backend.filter_boxes(proposals, 1)
-
-    proposals = keras.backend.gather(proposals, indicies)
-    scores = scores[:, :, :, :9]
-    scores = keras.backend.reshape(scores, (-1, 1))
-    scores = keras.backend.gather(scores, indicies)
-    scores = keras.backend.flatten(scores)
-
-    proposals = keras.backend.cast(proposals, tensorflow.float32)
-    scores = keras.backend.cast(scores, tensorflow.float32)
-
-    indicies = keras_rcnn.backend.non_maximum_suppression(proposals, scores,
-                                                          maximum, 0.7)
-
-    proposals = keras.backend.gather(proposals, indicies)
-
-    return keras.backend.expand_dims(proposals, 0)
 
 
 def resize_images(images, shape):
